Validations for Password and Email Added
Created validation methods for a user's password and email for
registration purposes.  The functions that the GUI will need to access
are CheckValidPassword and CheckValidEmailAddress.  We may need to
consider changing the CG_validator so that it doesn't directly attempt
to change the error label's text, but perhaps instead by passing the
text into the functions themselves from the GUI.  This will help loosen
the ties between the GUI layer and the Business layer.  Finally, we may
want to consider making a database manager for the business layer to
communicate through to access the database.  Just an idea to consider;
we'll figure it out.