Added CG_user and updated / document dbManager
Added the CG_user class with documentation and updated the dbManager
class to include documentation.

Updated main to contain a CG_dbManager and CG_user instance for the
persistence and business layers respectively. This is so that we can
pass them to any of the other widgets rather than just login. (say when
we have a lobby and or the game widgets)