Added in functionality for Tilesheets with CG_piece
Added functionality for tilesheets to be accepted and animated by
CG_piece, as well as a function to sync with Aaron's business layer once
pieces are added to CG_board.cpp's m_board. Temporarily they are all set
to a rook on every square, but they will be fed by the business layer
once it's complete.