Fixed Draggable Movements (Still lacking 1 thing)
Draggable movements now move more accurately and the piece being dragged
is now centered over the mouse for easier visibility.

This is still lacking the original movement method of being able to
click to pick up a piece and then click to place it, but I'll be adding
this back in and cleaning up alot more now that it's working better.

I also started to change the layout to give us more board space,
however, I realize that there's some funky things that happen with the
draw and resign buttons and the layout isn't quite finished but it
should be relatively soon.